Add 3 overload addTask function in Bot class to add Todo task, deadline task and event task.
This commit introduce a total of 3 overloaded addTask function in Bot class that creates Todo task, Deadline task or Event task based on the parameters given when adding a task. This creates a layer of abstraction for main class where correct type of Task will be created and added with matching parameters and managed by the Bot class.

This function added allows user to create task from interaction with Bot class.