Add exception handling for invalid command when formating user input
The program crashes when user input does not match a given format.

This commit implements exception handling of incorrect command format entered by the user and print feedback message with the correct command format.

The feedback message will update user of the correct format which allows user to know the exact correct command to enter instead of guessing for the correct format.

This reduces frustration of not knowing how to resolve the issue on user side, improving user experience and ensure program robustness.